From Mattias Ellert: If the TGFontTypeComboBox constructor detects that the requested X11 font is not installed and falls back to TGTextLBEntry::GetDefaultFontStruct(), then the destructor must not call gVirtualX->DeleteFont() on this font. The current version lacks protection for this which leads crashes when the default font struct is used after it has been wrongfully deleted by the TGFontTypeComboBox destructor. The backtrace generated can piont anywhere where the default font struct is used, but there is also an error message that says: Root > Error in <RootX11ErrorHandler>: BadFont (invalid Font parameter) (XID: 71303382, XREQ: 46) https://savannah.cern.ch/bugs/?75281
git-svn-id: http://root.cern.ch/svn/root/trunk@36649 27541ba8-7e3a-0410-8455-c3a389f83636

+++ gui/gui/src/TGComboBox.cxx
@@ -785,6 +785,7 @@ TGFontTypeComboBox::~TGFontTypeComboBox()
 {
    // Text font combo box dtor.
 
-   for (int i = 0; i < kMaxFonts && fFonts[i] != 0; i++)
-      gVirtualX->DeleteFont(fFonts[i]);
+   for (int i = 0; i < kMaxFonts && fFonts[i] != 0; i++) {
+      if (fFonts[i] != TGTextLBEntry::GetDefaultFontStruct()) gVirtualX->DeleteFont(fFonts[i]);
+   }
 }
